attributes: move destructuring tests to their own file
